did corrections

+++ src/main/java/com/riccardo/qr/EncodedBlock.java
@@ -6,7 +6,7 @@ class EncodedBlock {
     public final List<Integer> indices;
     public final int k;
     public final int totalBytes;
-    public final int checksum;   // or use CRC32
+    public final long checksum;   // or use CRC32
     public final byte[] data;    // The XORed block data
 
     public EncodedBlock(List<Integer> indices, int k, int totalBytes, int checksum, byte[] data) {
